registratino bug
