+++ frontend/src/features/auth/authSlice.js
@@ -1,5 +1,6 @@
 import { createSlice, createAsyncThunk } from "@reduxjs/toolkit";
 import axios, { setAuthHeader } from "../../api/axiosInstance";
+import { loginUser, registerUser } from "./authActions";
 
 // Load from localStorage
 const userFromStorage = JSON.parse(localStorage.getItem("user"));
@@ -11,53 +12,53 @@ if (tokenFromStorage) {
 }
 
 // Register User
-export const registerUser = createAsyncThunk(
-    "auth/register",
-    async (userData, thunkAPI) => {
-        try {
-            console.log("Sending POST /auth/register with:", userData);
-            const res = await axios.post("/auth/register", userData);
-            console.log("Register success:", res.data);
-            return res.data;
-        } catch (error) {
-            return thunkAPI.rejectWithValue(
-                error.response?.data?.message || "Register failed"
-            );
-        }
-    }
-);
+// export const registerUser = createAsyncThunk(
+//     "auth/register",
+//     async (userData, thunkAPI) => {
+//         try {
+//             console.log("Sending POST /auth/register with:", userData);
+//             const res = await axios.post("/auth/register", userData);
+//             console.log("Register success:", res.data);
+//             return res.data;
+//         } catch (error) {
+//             return thunkAPI.rejectWithValue(
+//                 error.response?.data?.message || "Register failed"
+//             );
+//         }
+//     }
+// );
 
 // Login User
-export const loginUser = createAsyncThunk(
-    "auth/login",
-    async (credentials, thunkAPI) => {
-        try {
-            const res = await axios.post("/auth/login", credentials);
+// export const loginUser = createAsyncThunk(
+//     "auth/login",
+//     async (credentials, thunkAPI) => {
+//         try {
+//             const res = await axios.post("/auth/login", credentials);
             
-            const { accessToken, user } = res.data;
+//             const { accessToken, user } = res.data;
 
-            // Save both in localStorage
-            localStorage.setItem("accessToken", accessToken);
-            localStorage.setItem("user", JSON.stringify(user));
+//             // Save both in localStorage
+//             localStorage.setItem("accessToken", accessToken);
+//             localStorage.setItem("user", JSON.stringify(user));
 
-            // Set token globally in axios
-            setAuthHeader(accessToken);
+//             // Set token globally in axios
+//             setAuthHeader(accessToken);
 
-            return { user };
-        } catch (error) {
-            return thunkAPI.rejectWithValue(
-                error.response?.data?.message || "Login failed"
-            );
-        }
-    }
-);
+//             return { user };
+//         } catch (error) {
+//             return thunkAPI.rejectWithValue(
+//                 error.response?.data?.message || "Login failed"
+//             );
+//         }
+//     }
+// );
 
-// Logout
-export const logoutUser = createAsyncThunk("auth/logout", async () => {
-    await axios.post("/auth/logout");
-    localStorage.removeItem("accessToken");
-    localStorage.removeItem("user");
-});
+// // Logout
+// export const logoutUser = createAsyncThunk("auth/logout", async () => {
+//     await axios.post("/auth/logout");
+//     localStorage.removeItem("accessToken");
+//     localStorage.removeItem("user");
+// });
 
 const authSlice = createSlice({
     name: "auth",
@@ -65,6 +66,7 @@ const authSlice = createSlice({
         user: userFromStorage || null,
         isLoading: false,
         error: null,
+        success: false,
     },
     reducers: {
         logout: (state) => {
@@ -83,14 +85,12 @@ const authSlice = createSlice({
                 state.isLoading = true;
                 state.error = null;
             })
+
             .addCase(registerUser.fulfilled, (state, action) => {
                 state.isLoading = false;
-                state.user = action.payload.user;
-                // localStorage.setItem("user", JSON.stringify(action.payload.user));
-                localStorage.setItem("accessToken", action.payload.accessToken);
-                localStorage.setItem("user", JSON.stringify(action.payload.user));
-                setAuthHeader(action.payload.accessToken);
+                state.success = true;
             })
+
             .addCase(registerUser.rejected, (state, action) => {
                 state.isLoading = false;
                 state.error = action.payload;
@@ -114,3 +114,122 @@ const authSlice = createSlice({
 
 export const { logout, clearError } = authSlice.actions; 
 export default authSlice.reducer;
+
+
+
+// import { createSlice, createAsyncThunk } from "@reduxjs/toolkit";
+// import axios, { setAuthHeader } from "../../api/axiosInstance";
+
+// // Load from localStorage
+// const userFromStorage = JSON.parse(localStorage.getItem("user"));
+// const tokenFromStorage = localStorage.getItem("accessToken");
+
+// // Set auth header if token exists on app load
+// if (tokenFromStorage) {
+//     setAuthHeader(tokenFromStorage);
+// }
+
+// // Register User
+// export const registerUser = createAsyncThunk(
+//     "auth/register",
+//     async (userData, thunkAPI) => {
+//         try {
+//             console.log("Sending POST /auth/register with:", userData);
+//             const res = await axios.post("/auth/register", userData);
+//             console.log("Register success:", res.data);
+//             return res.data;
+//         } catch (error) {
+//             return thunkAPI.rejectWithValue(
+//                 error.response?.data?.message || "Register failed"
+//             );
+//         }
+//     }
+// );
+
+// // Login User
+// export const loginUser = createAsyncThunk(
+//     "auth/login",
+//     async (credentials, thunkAPI) => {
+//         try {
+//             const res = await axios.post("/auth/login", credentials);
+            
+//             const { accessToken, user } = res.data;
+
+//             // Save both in localStorage
+//             localStorage.setItem("accessToken", accessToken);
+//             localStorage.setItem("user", JSON.stringify(user));
+
+//             // Set token globally in axios
+//             setAuthHeader(accessToken);
+
+//             return { user };
+//         } catch (error) {
+//             return thunkAPI.rejectWithValue(
+//                 error.response?.data?.message || "Login failed"
+//             );
+//         }
+//     }
+// );
+
+// // Logout
+// export const logoutUser = createAsyncThunk("auth/logout", async () => {
+//     await axios.post("/auth/logout");
+//     localStorage.removeItem("accessToken");
+//     localStorage.removeItem("user");
+// });
+
+// const authSlice = createSlice({
+//     name: "auth",
+//     initialState: {
+//         user: userFromStorage || null,
+//         isLoading: false,
+//         error: null,
+//     },
+//     reducers: {
+//         logout: (state) => {
+//             state.user = null;
+//             localStorage.removeItem("accessToken");
+//             localStorage.removeItem("user");
+//         },
+//         clearError: (state) => {
+//             state.error = null;
+//         },
+//     },
+//     extraReducers: (builder) => {
+//         builder
+//             // Register
+//             .addCase(registerUser.pending, (state) => {
+//                 state.isLoading = true;
+//                 state.error = null;
+//             })
+//             .addCase(registerUser.fulfilled, (state, action) => {
+//                 state.isLoading = false;
+//                 state.user = action.payload.user;
+//                 // localStorage.setItem("user", JSON.stringify(action.payload.user));
+//                 localStorage.setItem("accessToken", action.payload.accessToken);
+//                 localStorage.setItem("user", JSON.stringify(action.payload.user));
+//                 setAuthHeader(action.payload.accessToken);
+//             })
+//             .addCase(registerUser.rejected, (state, action) => {
+//                 state.isLoading = false;
+//                 state.error = action.payload;
+//             })
+
+//             // Login
+//             .addCase(loginUser.pending, (state) => {
+//                 state.isLoading = true;
+//                 state.error = null;
+//             })
+//             .addCase(loginUser.fulfilled, (state, action) => {
+//                 state.isLoading = false;
+//                 state.user = action.payload.user;
+//             })
+//             .addCase(loginUser.rejected, (state, action) => {
+//                 state.isLoading = false;
+//                 state.error = action.payload;
+//             });
+//     },
+// });
+
+// export const { logout, clearError } = authSlice.actions; 
+// export default authSlice.reducer;
